+++ src/SimulationTypes/Simulation_PowerRatioThreshold.cpp
@@ -57,7 +57,7 @@ void Simulation_PowerRatioThreshold::run()
     extern bool parallelism_enabled;
     #pragma omp parallel for ordered schedule(dynamic) if(parallelism_enabled)
 
-    for (unsigned i = 0; i < simulations.size(); i++)
+    for (size_t i = 0; i < simulations.size(); i++)
         {
         if (!simulations[i]->hasSimulated)
             {
