+++ src/SimulationTypes/Simulation_RegeneratorNumber.cpp
@@ -45,7 +45,7 @@ void Simulation_RegeneratorNumber::run()
     extern bool parallelism_enabled;
     #pragma omp parallel for ordered schedule(dynamic) if(parallelism_enabled)
 
-    for (unsigned i = 0; i < simulations.size(); i++)
+    for (size_t i = 0; i < simulations.size(); i++)
         {
         if (!simulations[i]->hasSimulated)
             {
@@ -429,8 +429,7 @@ void Simulation_RegeneratorNumber::createSimulations()
 {
     placeRegenerators(T);
 
-    for (unsigned long nreg = minRegNumber; nreg <= maxRegNumber;
-            nreg += stepRegNumber)
+    for (size_t nreg = minRegNumber; nreg <= maxRegNumber; nreg += stepRegNumber)
         {
         //Creates a copy of the Topology
         std::shared_ptr<Topology> TopologyCopy(new Topology(*T));
